+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+### v0.5.9 / 2024.10.07
+
+- feat: cache generated TTS file
+- feat: custom script and more #30
+  - feat: add custom i18n in config
+  - feat: add "Use rss items subtitle as title"
+- fix: use ttsCachePath
+
 ### v0.5.7 / 2024.10.06
 
 - fix: coqui-tts error on windows